requirements-web.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
piper-tts>=1.2.0
faster-whisper>=1.0.0
numpy>=1.24.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
